Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  83027.52560984132
  76424.66889138789
 107838.79208930467
  77626.48708805292
 108551.56480103644
  92266.81462239327
  87102.17164417623
 102228.76420421466
  94199.72361917849
 103443.61570539302
      ⋮
  82645.57284760838
  88781.51559063773
 107555.99010554806
  87707.05927811476
  93942.91165347854
 113131.21375001085
  84963.40948305847
 107189.25909856617
  72200.89586542945

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
  -84581.68028061782
 -109560.20422834685
  130541.06900115003
 -105725.76556764278
  101612.54987804662
  -70099.1108512299
  -79912.0936537591
  -13431.514512082387
   77117.90655184526
  160271.9104288088
       ⋮
  252014.4130187906
  149675.3020527931
    1744.6117506166029
  121736.4966067798
  -23301.02719819643
   36622.09453560911
  -86597.49207111553
   52370.41714854298
   60266.13276809483

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
  -23151.74721202209 - 20833.14587336652im
 -2704.6748941416745 - 7741.02179908776im
  -18403.39541444584 - 17041.624182432468im
  -33173.78183642227 - 19165.29621868356im
 -21803.529264595734 - 18147.17335565039im
  2351.2732688658034 - 5727.320275294849im
 -13088.874035096345 - 12058.255587235493im
 -18109.824906879156 - 22315.715815335243im
 -3886.1288981584803 - 8681.24144917315im
  -5347.764693288805 - 10547.479761746414im
                     ⋮
 -21970.074500439034 - 19141.840421471403im
  -63489.42839636738 - 26132.914127735046im
  -58117.39340012505 - 24912.73544883226im
  -33958.89980377718 - 23662.015904819636im
 -14485.394863823807 - 16699.63839097081im
  -36547.26659333281 - 22738.68534594713im
  3072.3481278924437 - 2476.3050059162824im
 -24791.008390838317 - 15919.788147398605im
  -5405.299454273234 - 9884.055200022587im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
  34795.134199080014 - 69132.37368166864im
   217545.9019395231 + 119337.26825571113im
  255279.70983020362 + 66317.05596287324im
 -22405.729563107983 + 29885.981732142758im
   246806.1759348074 + 57610.26565945762im
 -46055.667004874806 - 18707.00141739055im
  -96208.88180925525 - 30462.621604057207im
 -130988.89895004813 - 5254.873938913624im
  102765.55021989706 - 7184.968331896896im
 -39778.906740887964 - 139072.58146053206im
                     ⋮
  -171925.1998483347 - 29057.118815015034im
  -97947.24265665551 - 6124.210661268675im
   21883.99958270305 - 10710.547252580813im
 -177718.28193518554 - 18617.530032530492im
 -250207.46625429965 - 89195.06244205701im
 -340765.08522102446 - 101773.24871182353im
  -59431.21217847702 - 54987.88156124594im
  -96033.30244132064 - 20535.341630771698im
   43531.81436440942 - 18009.708806741324im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 73804.5  70788.6  59199.1  68189.6  …  66148.3  71059.9  56174.4  45948.8
 69002.8  72737.6  59666.4  66750.4     70612.7  75226.5  66687.2  45986.4
 74029.0  72847.1  56032.2  69192.1     70170.5  72893.4  58908.5  37736.0

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 73804.5  70788.6  59199.1  68189.6  …  66148.3  71059.9  56174.4  45948.8
 69002.8  72737.6  59666.4  66750.4     70612.7  75226.5  66687.2  45986.4
 74029.0  72847.1  56032.2  69192.1     70170.5  72893.4  58908.5  37736.0