Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  95750.49425691017
  95749.14149674625
  80244.7868777895
 102381.134338342
  98457.92812919378
  81843.23024730159
  94829.39034016522
  98990.06956318967
 108296.04541830017
 107515.52587499916
      ⋮
  75445.63256558018
  88195.56263245353
  96647.24462284843
 103722.61604586006
  99564.92400266754
  88933.61153906232
  90183.31474396397
  89900.3948269757
  80261.15988016964

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
   83321.72106518278
  102831.44109935421
  113241.23509731633
  198088.16367250634
  -27378.18591520457
  256741.0745694326
  -24056.279287568956
  180044.18801110963
  146132.5862744975
  134201.2726879282
       ⋮
 -108303.30336306151
  -73626.3062996639
  -27154.185575168787
   19569.635033830084
  226221.5239701569
  206061.532517463
  -79100.8806491776
  216298.42778985374
  197527.32719479746

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
 -14465.750847201965 - 11030.213141978751im
 -15288.930814663177 - 16016.051853935283im
  -34970.52770179759 - 18195.81934370202im
  -18316.24993172978 - 14978.01782986599im
  -15045.58067299744 - 16571.273111761395im
    4764.86906996964 - 4166.949395175622im
 -17887.700468852276 - 12583.166942554435im
   8979.444367134074 + 931.7065695938809im
 -19729.424274754154 - 20374.2721181771im
  2520.4691717287283 - 6445.009792834929im
                     ⋮
  -79596.55204095796 - 29078.295616881318im
   7355.795842040021 - 1885.3831426688114im
 -32158.535205910586 - 23424.883246789726im
  -39842.98986890091 - 21494.250458474435im
  3350.2283507029397 - 6929.531103145255im
 -1800.1470774903864 - 9175.727845119322im
    8522.68463606125 + 481.4992180657971im
 -15126.314017647619 - 17850.555231390394im
  -6270.159638605262 - 11257.845517158446im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
  -70432.47644906308 - 65143.23027100302im
   80006.37095768758 - 26028.7671325567im
 -223576.07328316744 - 60645.2319820042im
  -4193.625779393762 + 36827.440156418714im
 -39321.479837137056 + 33759.064936242765im
  43037.323454652236 - 2212.517533925391im
  308754.81709147315 + 110541.84116641841im
   -9628.16540855444 - 69202.27140112307im
 -114296.29336782424 - 2863.6291222240397im
  125285.17897759208 + 75661.29368243882im
                     ⋮
 -121778.35612870274 - 83084.62899037851im
  21870.953556506298 + 30168.554114536557im
 -221524.61851682255 - 37076.06756109616im
  3130.8172311351245 - 50251.394480200936im
  10045.520055702422 + 199.21424386181752im
   70324.02502201845 + 92739.87671603505im
  -9105.984903868964 - 29148.453944323206im
 -13529.436924322461 + 56849.52818628673im
 -1932.9719829028854 + 16743.179401692956im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 68217.8  61882.1  57299.7  72746.9  …  52685.3  67304.2  52563.6  56951.8
 65823.3  57877.6  57169.9  73084.5     43351.3  64119.3  53924.7  45704.9
 63645.8  59909.4  47625.1  74285.9     55731.1  65900.3  53800.4  56257.3

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 68217.8  61882.1  57299.7  72746.9  …  52685.3  67304.2  52563.6  56951.8
 65823.3  57877.6  57169.9  73084.5     43351.3  64119.3  53924.7  45704.9
 63645.8  59909.4  47625.1  74285.9     55731.1  65900.3  53800.4  56257.3