Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  90182.81883769753
  72918.97114647784
  96869.29923292901
  94309.03810576045
  85498.4514906978
  98837.80633647858
  75160.47767430736
  82916.8876472834
 102724.22932101385
  86109.65442179581
      ⋮
  91675.81673683466
  88236.95049485336
  97259.33242386847
 105343.0431139115
  90715.76333452512
  76344.73402105222
 110199.06813215662
  91361.50143868239
  99754.89195786035

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
 173722.99670380133
  -9497.637195156814
 229360.46084562087
 139657.57651343636
 220902.80668004084
 160192.0904235573
  53012.821929630925
 -88607.44706479413
  -3973.1334055981843
 -90089.53415168142
      ⋮
 150288.3341302185
 197930.64527266892
 -41190.16986239308
 207011.15526442855
 254964.28303506732
  86236.65025735353
  77637.66452443184
 -50158.15300199546
 161780.72820309422

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
  -15027.92421261138 - 12349.25240943391im
  -19650.46786165136 - 16137.670477882337im
 -18911.592831547558 - 17799.399222773995im
 -1509.8040104195961 - 9386.424182899427im
 -33889.928771462146 - 18634.493995980953im
  -26016.33665460605 - 17367.16327885153im
  -55317.24217532646 - 23444.864711654518im
 -29732.232398404183 - 19939.889915433803im
   5771.583675571639 - 1896.738869380165im
 -29145.785221322752 - 21039.406520377073im
                     ⋮
 -19048.092309225336 - 16667.740306160045im
   4639.310218331261 - 4926.34846547402im
 -10291.478700683769 - 10154.350294028742im
  1205.2257321494258 - 5135.159372004453im
 -48365.002882327986 - 24196.633687272795im
   -1862.18213755344 - 6321.905048536694im
 -38831.003143898415 - 18592.271961065333im
   -34949.5741634795 - 20715.68704990296im
  -30686.12666549147 - 17547.094673650652im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
   62224.05811729245 + 25375.688191641355im
   -30902.8102448341 + 24817.160586116734im
 -29969.513371559493 - 77630.1108746344im
  -30541.29246584905 - 1050.6970429420453im
  252038.12780379955 + 52566.205246738195im
 -252248.87694447913 - 79904.67384885733im
   -412995.717440427 - 121463.63449100929im
  -325128.6849869783 - 101073.73890980249im
  63008.265315644996 + 79488.07043676346im
   70532.10825017176 - 25810.335944681712im
                     ⋮
  -7729.799380748827 - 35161.86676124434im
  20487.908840464344 + 31220.54388181886im
  -115880.4405066384 - 20451.276601262438im
    50271.7155265715 + 33453.051877347614im
   80594.65185766076 - 8237.902639253633im
   66015.94409670186 + 70929.33037928396im
 -150622.53583524717 - 53072.77271742608im
  -82197.10507184124 + 8683.002468547998im
  11726.500405661453 + 25209.41367895464im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 71663.6  66560.8  55004.9  73404.3  …  62553.3  65548.3  73871.1  61882.7
 70858.1  66765.0  68094.7  68821.7     52319.0  63529.0  73956.0  55571.1
 70117.5  59258.2  63989.2  71873.4     62886.2  65681.8  76234.5  61736.4

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 71663.6  66560.8  55004.9  73404.3  …  62553.3  65548.3  73871.1  61882.7
 70858.1  66765.0  68094.7  68821.7     52319.0  63529.0  73956.0  55571.1
 70117.5  59258.2  63989.2  71873.4     62886.2  65681.8  76234.5  61736.4