Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  84877.43050533588
 105688.84344937622
  87157.7560605252
  92716.81606481466
  91208.96116973067
 102002.22461161914
  75021.33501181887
  70233.92662208964
 117591.076713744
  90612.61061906288
      ⋮
 107420.50239023835
  92399.56755635657
 112947.3676015426
 111463.20504001644
  85886.97530098897
  90796.38288217099
  84539.06209518896
  81910.20628461949
  79982.32605363971

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
  64414.99368308596
 214036.65328176718
 -49255.15859625961
 173000.50616688104
  39513.403247975104
 216368.6919212193
 175778.37901377454
  80168.37390770494
  88194.081429204
 140296.90892799362
      ⋮
 140672.95763951092
 178353.5834308641
 127629.65941193428
  94323.01843577909
 -83892.30480542236
 -76944.60590378434
  -9515.814902600127
  92171.69139086959
 101189.12434396624

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
  -38801.44447082009 - 21475.97965703091im
 -11582.586391532932 - 13820.109905826477im
 -17950.679677922646 - 14888.765801801812im
  -83229.09467831474 - 30232.820434743615im
   532.6027054038068 - 6919.047468897386im
  -71651.25187377392 - 27418.711460848557im
   4105.209180592794 - 3696.810230725395im
  -69475.52375939664 - 27306.81986172044im
 -22537.705310636917 - 17916.81738958683im
  -61868.96664373603 - 25212.87511617763im
                     ⋮
  -802.0187478848798 - 8751.842473149782im
  -56032.85091375664 - 23719.599495093204im
  1150.3762221220732 - 8819.5738175991im
  -8483.710242472443 - 11522.321862041317im
    9088.54324137303 + 3194.165235954556im
  2410.3796042935064 - 7259.6539021123435im
   7513.494399815243 - 72.54705125759415im
  -29860.19064095979 - 21474.686992230556im
  -42492.51435620513 - 21245.03357329316im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
   277766.2744800325 + 46549.62461622922im
   33158.00420649684 - 11983.038660568185im
   54800.20123018273 + 1791.8190403852632im
   7262.581831426127 + 1176.681331320986im
 -151749.56934685103 - 93959.82235802953im
 -176351.17489900393 - 50391.59591596124im
  -38034.12846799327 - 45821.80959948378im
  117173.13137537125 - 8291.975212263584im
   -275638.532848765 - 84462.53375393979im
 -117987.54680874254 - 62811.506162219004im
                     ⋮
   46948.53604352385 - 27150.108837739608im
 -143716.51900725183 - 42618.33461469779im
 -127838.80089401003 - 70330.34033556483im
   40294.39527667844 + 69855.36088257044im
  13681.486657284295 + 21830.237912068747im
  -98030.91889958293 - 60488.72907411924im
   17292.86870504591 + 10485.638266174814im
 -125456.07040220969 + 26383.0073195344im
  124606.56418615824 + 22181.318208300538im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 40204.6  71058.3  50783.3  70685.9  …  52613.3  64784.2  54917.6  62561.8
 53055.1  64539.6  59567.3  65843.5     52243.6  60296.0  56589.1  70441.5
 55272.1  71407.3  63095.0  70207.9     49558.2  70640.4  48755.0  69912.8

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 40204.6  71058.3  50783.3  70685.9  …  52613.3  64784.2  54917.6  62561.8
 53055.1  64539.6  59567.3  65843.5     52243.6  60296.0  56589.1  70441.5
 55272.1  71407.3  63095.0  70207.9     49558.2  70640.4  48755.0  69912.8