Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  73465.50913689815
  79009.94727048583
  97011.90202019706
 100634.52006703799
  78376.11734646474
 101644.56500182727
 117187.84450792456
  95124.07990714224
  87331.27317989108
 102866.20644935056
      ⋮
 108188.04755766992
  98156.2061475169
  88115.45768945987
  92539.17176154266
  99708.70998689416
 102259.13826940193
  84090.06601407958
  78709.56396799565
  74679.65532233758

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
 -107522.10757553668
   36036.24247271304
  114395.04459501371
   41831.61425382557
  -69576.07522426189
  207838.89600483442
  152503.43782100614
  -24358.734609531013
  232521.80845337675
  -12443.051665110943
       ⋮
     931.4562317123491
   23423.375871778415
  142206.8486924446
  -40924.65216577622
   38674.599565087774
  161034.9369754904
   51196.7245546057
  243025.4488081946
  258673.0945154358

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
  -61153.31159877791 - 27133.43006563417im
 -15403.262987521774 - 16837.689667920135im
   7658.543224509228 - 73.99468168121075im
  -51689.75513585431 - 24492.917915944214im
 -12260.577616792081 - 12083.309366361853im
  -2502.597736582242 - 6053.065916968877im
  -68651.27713023103 - 27988.81908490642im
 -51151.747260592034 - 23448.586450367617im
  -76042.56647544862 - 29911.720992780945im
  1371.8459750883076 - 7625.831324657709im
                     ⋮
 -31520.286696712526 - 17889.25238156403im
 -23210.779058435877 - 18743.13080406552im
  -8722.406540155735 - 14993.214248489801im
  -46491.30271543168 - 22201.54972364334im
   8818.179126175275 + 1515.5827951178019im
  -72893.97776148022 - 28758.928028213413im
  -41954.44671407678 - 23014.474452091275im
 -32257.211246966268 - 18378.30730345577im
 -31573.560525648198 - 20498.48787890942im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
  135525.62990144273 + 11814.882271662249im
 -130280.11544320424 - 10116.789128231896im
 -27043.552972385227 - 31489.63166259674im
  -325924.4151245932 - 91583.96486004136im
  139410.92110083342 + 17728.401364112207im
    89085.5175035324 + 17815.11661407433im
 -204575.29186480195 - 43049.23243937831im
 -166865.72432539915 - 47094.57346365397im
 -207243.06307524152 - 45581.74034289482im
  127854.46547410496 + 60774.5470200031im
                     ⋮
  311455.19233030616 + 86777.32937156857im
  -264678.6315740902 - 78822.66523093788im
   49606.53734868729 - 52406.11717185047im
 -28399.219399832637 - 32501.34363364106im
   -8189.80069527428 - 70478.10013511573im
  -473055.6897924682 - 192208.8186172221im
 -281214.77609303227 - 77134.07875015948im
  -5571.771352321601 + 7785.431975405925im
   -68116.8906537724 + 32228.506743337595im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 66538.7  67006.1  69676.7  71501.2  …  59473.9  61385.4  66620.1  68347.5
 65992.6  67313.7  71745.4  71086.3     60669.8  61993.4  59838.3  73982.8
 72579.4  59118.3  70469.0  73765.8     59612.1  58370.1  54319.6  71593.1

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 66538.7  67006.1  69676.7  71501.2  …  59473.9  61385.4  66620.1  68347.5
 65992.6  67313.7  71745.4  71086.3     60669.8  61993.4  59838.3  73982.8
 72579.4  59118.3  70469.0  73765.8     59612.1  58370.1  54319.6  71593.1