Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  93302.45332073147
  92950.00337611715
 103054.16041127376
  98132.72014898044
 115816.36064216161
  75272.94468243365
  91016.57635588411
 101307.98955877042
  96723.98720047259
  76357.54421723586
      ⋮
 103141.42503272733
  97757.79808413141
 102627.27175180287
 103871.45868555624
  83669.48772521924
  88638.57948686702
  90583.88476049682
 103414.45484848398
  76547.80117594454

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
 -58799.69339991592
 234779.14728356356
 130429.72416147494
 154920.74808912002
 157955.63582216532
   9598.360927055804
 131186.75685449166
  21187.334849012208
 146078.81054629435
 176819.01930715758
      ⋮
 219194.1137865559
 163534.240634954
  42120.2172301309
 150308.21657302356
 221471.27129988716
 235181.9186420761
 -54483.39493629983
 140489.0290884657
  90749.67634475115

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
 -19601.069103008726 - 14620.721362855416im
   3561.999216507923 - 3057.637287960132im
  1606.2256549187364 - 6067.96803201751im
   951.4206348183379 - 8503.416176846364im
 -15482.527166575179 - 12979.740273965559im
   -6018.14688212571 - 10991.62142470744im
 -12165.465241240756 - 12535.630709021561im
  -13455.65829045759 - 15346.666200925298im
    3966.97312526754 - 4631.394101079771im
 -18190.382561843428 - 19125.96529989739im
                     ⋮
  -6841.736934408293 - 12253.06503765771im
 -22943.544937824197 - 22304.393617575526im
  -51187.62739326217 - 24880.023266580487im
  5211.9970199405925 - 3813.1739181042135im
 -21333.150316788473 - 20140.0659905579im
   2074.355097169615 - 5637.0249367797205im
   844.5729768382562 - 8852.876465412039im
  -4140.412417502992 - 7006.638740737472im
 -25125.920655830087 - 21910.207442339113im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
 -148699.89807880696 - 38118.611652119886im
  65062.148027558054 + 14703.263747669645im
  -26891.86452830657 + 13010.645458157407im
  -25324.79090506725 - 52500.50795526398im
  -63071.49775180941 - 75766.20324459024im
  -7692.063431337656 + 4280.858192174013im
   32147.94850688832 + 5678.0124917548965im
  -72068.39304474833 + 1904.4824893862897im
   43773.57018761031 - 8110.679279889744im
  52224.648743399695 - 47211.0963836484im
                     ⋮
 -1695.3458239375723 - 43527.55280993767im
 -204881.40319227165 - 39324.21786489203im
  103253.15336681729 + 3273.887166912802im
  34608.301507437296 + 33253.25958254664im
  -269613.4884449641 - 86961.88952157945im
  37402.535520242374 - 9420.385410737974im
 -3375.1872264765925 - 21245.47474007417im
  -55221.44739274628 + 2993.5951420714264im
    94159.5717913176 - 32354.279138519032im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 63414.8  78827.1  55548.2  74199.3  …  52875.4  61352.8  58276.3  72325.6
 63697.6  78461.9  61931.8  74154.2     55073.4  71449.7  58165.7  66696.5
 70924.3  78578.3  61838.2  71200.3     54798.8  65017.2  51271.2  69412.8

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 63414.8  78827.1  55548.2  74199.3  …  52875.4  61352.8  58276.3  72325.6
 63697.6  78461.9  61931.8  74154.2     55073.4  71449.7  58165.7  66696.5
 70924.3  78578.3  61838.2  71200.3     54798.8  65017.2  51271.2  69412.8