Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  94174.89356614124
  91312.30549240485
  99084.46652905601
  73223.8157656373
 108015.73956489189
  98297.25234574621
  91523.50882962749
 113942.85335919331
 109548.52465459623
 112075.51500529758
      ⋮
  81418.91474382204
  96445.5642807159
  88782.07009062522
 113496.63662851174
  81224.83896386011
 100435.9904519798
  93486.17867336611
 104614.56126222109
 115046.28505960412

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
 -31102.61407438576
 229543.08966033976
 207918.7554374414
  71468.18222865247
  59274.521843418755
 180369.83729717348
 221254.95251344226
 133537.3981182936
  26277.821120397068
  90933.36623927546
      ⋮
 253198.06309745755
 154670.90676771902
 -20997.99513620486
 100499.85054302533
  96242.57221624744
  96285.74658458639
 216507.213158809
  -9587.358910904459
 166115.21404233782

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
  -76147.45787630488 - 29370.622323738095im
  2994.0954836116944 - 6500.244560594117im
  -9703.780842535441 - 9960.639343290179im
 -10889.862389224887 - 15851.444804519899im
  -1030.113788678469 - 9012.603933591483im
   8773.188181666077 + 1403.8706804272388im
   2588.073486709004 - 5473.564241171032im
 -17292.742260674597 - 14346.92418892053im
  -27041.43338275273 - 19036.441743928146im
   -44051.1705707404 - 23717.28875753092im
                     ⋮
  -36870.71472595996 - 19848.31496071793im
  2309.1927906674373 - 4236.631719894195im
 -27504.887908481265 - 18476.556145826366im
  -2764.031045762036 - 9143.507771951441im
  -50746.51224371273 - 23242.32150378355im
  -6982.004103642215 - 13178.583289655686im
 -36339.136439282505 - 18840.008923563386im
  -8002.518248692767 - 14175.15779546817im
  -35647.90060411113 - 21114.294570050988im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
 -10980.008521265292 - 11356.248331039846im
 -19889.620674162285 - 33774.070986751074im
   29235.57108768032 + 18019.83624374918im
   139801.0156146478 + 12281.345643789782im
  143596.82373090752 + 55170.65294924767im
   49156.82182266504 + 70422.01891258724im
 -132428.36656299833 - 86087.76534150669im
  -74142.12841607089 - 11030.59772114615im
  -6358.864372517397 - 62834.63174359272im
 -203497.05942234595 - 32626.77364514126im
                     ⋮
 -32117.674474243322 + 5996.650373885728im
  -78691.31021007386 - 17384.0235589625im
  211970.60816348257 + 31836.595146515887im
  -46347.67810656819 - 76501.04916810509im
  -95027.80344880259 - 31522.771999851513im
   43438.26892969629 + 80791.72394030943im
 -125293.10157069273 - 62213.461581974676im
   51821.33293514386 + 89244.6222703382im
   7303.090299825737 - 50612.14924981325im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 53938.2  60728.3  64598.5  57259.8  …  74013.5  45625.0  74652.2  67840.7
 62569.4  53103.6  69308.2  66635.9     74308.0  29839.0  74554.3  70483.0
 61527.8  61649.4  72091.3  68437.9     74287.8  45216.6  73648.6  70677.4

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 53938.2  60728.3  64598.5  57259.8  …  74013.5  45625.0  74652.2  67840.7
 62569.4  53103.6  69308.2  66635.9     74308.0  29839.0  74554.3  70483.0
 61527.8  61649.4  72091.3  68437.9     74287.8  45216.6  73648.6  70677.4