Predefined kernels

While it is certainly possible to define your own kernel, doing so efficiently can be non-trivial and time-consuming. IFGF.jl provides a few predefined kernels that are commonly used in practice. If you need a kernel that is not listed here, or if you have implemented a custom kernel and would like for it to be added to the library, please open an issue and/or PR on the GitHub repository.

The kernel-specific API is heavily inspired by the FMM3D library, and provides functions to efficiently compute sums of the form

\[u(\boldsymbol{x}_i) = \sum_{j=1}^N G(\boldsymbol{x}_i - \boldsymbol{y}_j) c_j +\gamma_{1,\boldsymbol{y}} G(\boldsymbol{x}_i - \boldsymbol{y}_j) v_j,\]

where $\boldsymbol{x_i}$ are the target locations, $\boldsymbol{y}_j$ are the source locations, $G(\boldsymbol{x} - \boldsymbol{y})$ is (up to a constant factor) the fundamental solution of a given partial differential operator, $\gamma_{1,\boldsymbol{y}}$ its (generalized) Neumann trace, and $c_i,v_i$ are input vectors of the appropriate size. IFGF.jl computes the sum above in log-linear complexity.

Each predefined kernel comes with an associated plan function that precomputes the necessary data structure required to efficiently apply the forward map. If you need to compute the forward map multiple times with the same kernel, it is a good idea to build a plan.

In the following, we provide a brief overview of the predefined kernels.

Single-precision

All predefined kernels support single-precision arithmetic. If you do not need double-precision, you can save memory and computation time by using single-precision to represent your data.

Laplace 3D operator

IFGF.laplace3d — Function

laplace3d(targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} - \frac{(\boldsymbol{x}_i - \boldsymbol{y}_j)^\top v_j}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||^3} ,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{T}: vector of n charges.

dipvecs::Matrix{T}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf (e.g. tol)

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(m)
dipvecs = rand(3,m)
IFGF.laplace3d(targets, sources; charges, tol=1e-4)

100000-element Vector{Float64}:
  87899.27435572754
  98547.03902968366
  98269.67521427381
  98319.61703667634
  86777.62218655417
 100244.08487119748
  91417.95711762136
  84765.89897016347
  77128.76369015583
 105561.6223633966
      ⋮
 104135.47107131233
  91705.92863010586
 106553.68522569179
  86701.39130301015
  86006.57436476741
  96057.85540098071
  84423.32742209989
 106092.00524704087
 108180.50819536112

Plan and forward map:

L = IFGF.plan_laplace3d(targets, sources; charges, dipvecs, tol=1e-4)
out  = zero(charges)
IFGF.laplace3d!(out, L; charges, dipvecs)

100000-element Vector{Float64}:
  -3207.6060520757783
 -12390.95338854372
 237860.8742803534
   2609.1018185682533
  -8401.730737219255
 -26572.129384657735
 -24807.155521062887
 121097.22117061105
  -6364.7166512782205
 100538.10049313864
      ⋮
  71018.10489123654
 -16260.111275424506
 128463.37579348877
  67739.31516298326
 -36807.83990018486
 215330.73120520523
 -70618.85024661421
 192391.48956420444
 196297.8132842856

Helmholtz 3D operator

IFGF.helmholtz3d — Function

helmholtz3d(k, targets, sources; charges = nothing, dipvecs = nothing, grad =
false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \frac{c_j e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} + \nabla_{\boldsymbol{y}}^\top \left( \frac{e^{i k || \boldsymbol{x}_i - \boldsymbol{y}_j ||}}{|| \boldsymbol{x}_i - \boldsymbol{y}_j ||} \right) v_j,\]

where $c_j$ are the charges and v_j are the dipvecs.

Input:

k: the wavenumber.

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

charges::Vector{Complex{T}}: vector of n charges.

dipvecs::Matrix{Complex{T}}: 3 x m matrix of dipole vectors.

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: the potential at the target points if grad=false, or the gradient as a 3 x n matrix if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
charges = rand(ComplexF64,m)
dipvecs = rand(ComplexF64,3,m)
k = 2π
IFGF.helmholtz3d(k, targets, sources; charges, tol=1e-4)

100000-element Vector{ComplexF64}:
 -20488.135296792912 - 15449.413262315049im
  -7823.924308329492 - 12439.219670245493im
  -85463.25890806652 - 31050.901531520365im
  -6749.630031057217 - 13815.224849264578im
  -68053.78522857158 - 26962.354201509217im
 -47517.983460889605 - 23518.19350073142im
 -10681.716289286569 - 11843.014433642344im
   3663.120701538208 - 7197.425302767802im
  -8869.932926258865 - 13274.517711074568im
   9854.179523318977 + 3118.684252076751im
                     ⋮
 -32004.419180034325 - 19608.413865477243im
   9842.904046218926 + 3201.706775539842im
  -39398.53839291879 - 24051.578350013933im
 -14721.568366298967 - 14226.070885368597im
 -32568.221806345566 - 18635.56217456311im
  -58321.76949634232 - 25356.462835270955im
  -39156.13671723998 - 20498.67430496143im
 -2386.9990090028787 - 10621.866975014405im
 -27470.167669613875 - 19579.36590777498im

Plan and forward map:

L = IFGF.plan_helmholtz3d(k, targets, sources; charges, dipvecs, tol=1e-6)
out  = zero(charges)
IFGF.helmholtz3d!(out, L; charges, dipvecs)

100000-element Vector{ComplexF64}:
   74126.45418950915 + 1901.4090551036131im
  223224.73430287564 + 68986.10830788926im
 -32084.999344085685 - 9001.847339169923im
  -4679.551209098743 - 67593.48537623746im
   51485.72741321852 + 7028.405172615137im
 -130462.37110505953 + 2711.2032119645373im
  -7721.286274711996 + 23124.68522423961im
  -60102.32528766112 - 54166.636233273806im
 -177273.19497775572 - 58297.91982882595im
   52061.12435520683 + 34650.87845852937im
                     ⋮
   166468.3528604781 + 3015.053895739276im
  23477.194095145547 + 520.9906530400494im
  122543.38534601982 - 22114.784914509488im
   212804.7879459711 + 36475.22390355877im
 -287817.43573435314 - 89203.55051838167im
  124008.55089326476 - 1597.2638383573794im
  254028.11480919746 + 48552.6719393824im
 -146740.65871285484 - 52390.009858883706im
 -157584.03549428022 - 18733.36566271935im

Stokes 3D operator

Warning

Stokes 3D operator should be considered experimental. Please report any issues you encounter.

IFGF.stokes3d — Function

stokes3d(targets, sources; stoklet = nothing, strslet = nothing, strsvec =
nothing, grad = false, kwargs...)

Compute the sum

\[u(\boldsymbol{x}_i) = \sum_{j=1}^n \left( \frac{I}{d} + \frac{\boldsymbol{r}_{ij}\boldsymbol{r}_{ij}^\top}{2d_{ij}^3} \right) \boldsymbol{\sigma}_j + \frac{3(\boldsymbol{\nu}_j \cdot \boldsymbol{r}_{ij})(\boldsymbol{\mu}_j\cdot\boldsymbol{r}_{ij}) \boldsymbol{r}_{ij}}{d_{ij}^5},\]

where $\boldsymbol{r}_{ij} = |\boldsymbol{x}_i - \boldsymbol{y}_j|$, $d_{ij} = |\boldsymbol{r_{ij}}|$, $\boldsymbol{\sigma}_j$ are the Stokeslet strengths, $\boldsymbol{\mu}_j$ are the stresslet strengths, and $\boldsymbol{\nu}_j$ are the stresslet orientations.

Input:

targets::Matrix{T}: 3 x n matrix of target points.

sources::Matrix{T}: 3 x m matrix of source points.

stoklet::Matrix{T}: 3 × n matrix of stokeslets

strslet::Matrix{T}: 6 x m matrix of stresslet densities (indices 1:3) and orientations (indices 4:6)

grad::Bool: if true, the gradient is computed instead

kwargs...: additional keyword arguments passed to assemble_ifgf.

Output

u: 3 x n matrix giving the velocity at the target points if grad=false, or 3 x 3 × n matrix representing the gradient of u if grad=true.

Forward map:

using IFGF
m = 100_000
targets = sources = rand(3, m)
stoklet  = rand(3,m)
strslet = nothing
IFGF.stokes3d(targets, sources; stoklet, strslet, tol=1e-4)

3×100000 Matrix{Float64}:
 51009.4  69385.7  45554.7  50337.9  …  42535.1  67342.6  49590.9  57419.6
 58242.4  70737.3  44818.0  46954.8     54716.6  57920.3  44958.8  57466.8
 53895.1  68947.6  39991.4  47818.8     58753.7  60272.7  50590.7  55485.8

Plan and forward map:

L = IFGF.plan_stokes3d(targets, sources; stoklet, strslet, tol=1e-4)
out = zero(stoklet)
IFGF.stokes3d!(out, L; stoklet)

3×100000 Matrix{Float64}:
 51009.4  69385.7  45554.7  50337.9  …  42535.1  67342.6  49590.9  57419.6
 58242.4  70737.3  44818.0  46954.8     54716.6  57920.3  44958.8  57466.8
 53895.1  68947.6  39991.4  47818.8     58753.7  60272.7  50590.7  55485.8